# expression: implement vectorized evaluation for 'builtinNEIntSig'