--- a/expression/builtin_compare_vec.go
+++ b/expression/builtin_compare_vec.go
@@ -267,11 +267,35 @@ func (b *builtinEQIntSig) vecEvalInt(input *chunk.Chunk, result *chunk.Column) e
 }
 
 func (b *builtinNEIntSig) vectorized() bool {
-	return false
+	return true
 }
 
 func (b *builtinNEIntSig) vecEvalInt(input *chunk.Chunk, result *chunk.Column) error {
-	return errors.Errorf("not implemented")
+	n := input.NumRows()
+	var err error
+	var buf0, buf1 *chunk.Column
+	buf0, err = b.bufAllocator.get(types.ETInt, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(buf0)
+	if err := b.args[0].VecEvalInt(b.ctx, input, buf0); err != nil {
+		return err
+	}
+	buf1, err = b.bufAllocator.get(types.ETInt, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(buf1)
+	if err := b.args[1].VecEvalInt(b.ctx, input, buf1); err != nil {
+		return err
+	}
+
+	result.ResizeInt64(n, false)
+	vecCompareInt(mysql.HasUnsignedFlag(b.args[0].GetType().Flag), mysql.HasUnsignedFlag(b.args[1].GetType().Flag), buf0, buf1, result)
+	result.MergeNulls(buf0, buf1)
+	vecResOfNE(result.Int64s())
+	return nil
 }
 
 func (b *builtinGTIntSig) vectorized() bool {
@@ -425,6 +449,17 @@ func vecResOfLT(res []int64) {
 	n := len(res)
 	for i := 0; i < n; i++ {
 		if res[i] < 0 {
+			res[i] = 1
+		} else {
+			res[i] = 0
+		}
+	}
+}
+
+func vecResOfNE(res []int64) {
+	n := len(res)
+	for i := 0; i < n; i++ {
+		if res[i] != 0 {
 			res[i] = 1
 		} else {
 			res[i] = 0