--- a/expression/builtin_compare_vec.go
+++ b/expression/builtin_compare_vec.go
@@ -532,7 +532,7 @@ func vecResOfNE(res []int64) {
 	n := len(res)
 	for i := 0; i < n; i++ {
 		if res[i] != 0 {
-      res[i] = 1
+			res[i] = 1
 		} else {
 			res[i] = 0
 		}