<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for builtinNEIntSig, for #12103


### What is changed and how it works?
```
BenchmarkVectorizedBuiltinCompareFunc/builtinNEIntSig-VecBuiltinFunc-8         	  391096	      2875 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinNEIntSig-NonVecBuiltinFunc-8      	   28605	     42841 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinNEIntSig-VecBuiltinFunc#01-8      	  414004	      2751 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinNEIntSig-NonVecBuiltinFunc#01-8   	   29678	     41874 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinNEIntSig-VecBuiltinFunc#02-8      	  369595	      3242 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinNEIntSig-NonVecBuiltinFunc#02-8   	   28356	     41751 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinNEIntSig-VecBuiltinFunc#03-8      	  361756	      3357 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinNEIntSig-NonVecBuiltinFunc#03-8   	   24622	     47705 ns/op	       0 B/op	       0 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test